* kleine Änderungen bei der Ausgabe von ex07_01 (es fehlt noch "visualize the classification error in comparision ...") *Änderung des ersten classifiers in haarlikeFeatures (muss noch bei anderen classifiers angepasst werden, aber ich werd evtl morgen mal nachfragen)

+++ ex07/ex07_02.m
@@ -1,6 +1,6 @@
-img1 = rgb2gray(imread('./test_images/face1.jpg'));
-img2 = rgb2gray(imread('./test_images/face2.jpg'));
-img3 = rgb2gray(imread('./test_images/face3.jpg'));
+img1 = double(rgb2gray(imread('./test_images/face1.jpg')));
+img2 = double(rgb2gray(imread('./test_images/face2.jpg')));
+img3 = double(rgb2gray(imread('./test_images/face3.jpg')));
 
 intImg1 = constructIntegralImage(img1);
 % intImg2 = constructIntegralImage(img2);
